Back End/talk-harmony-api/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
blinker==1.6.2
cffi==1.15.1
click==8.1.3
dnspython==2.3.0
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.1
pycparser==2.21
PyJWT==2.7.0
pymongo==4.3.3
python-dotenv==1.0.0
Werkzeug==2.3.4
